refactor: extract remaining logic from JSX into hooks (separate-logic run 2)
Continues the logic/UI separation started in #112, clearing the last
three JSX components that still contained business logic.

New hooks created:
- useContactLegacyForm — form state + async submit (Contact.jsx)
- useAccountInit — reactive useEffect fetch on selectedCar change (Account.jsx)
- useAccountTableData — useCallback filter, handleCar handler, store selectors (AccountTable.jsx)
