+++ client/src/components/landing/contact/hooks/useContactLegacyForm.js
@@ -1,0 +1,13 @@
+import { useState } from "react";
+import { createContact } from "../../../../utils";
+
+export function useContactLegacyForm() {
+  const [formData, setFormData] = useState({ from: "", title: "", description: "" });
+
+  const onSubmit = async (e) => {
+    e.preventDefault();
+    await createContact(formData);
+  };
+
+  return { formData, setFormData, onSubmit };
+}
