+++ client/src/pages/account/Account.jsx
@@ -1,17 +1,10 @@
 import "./account.css";
-import { useEffect } from "react";
 import ReqService from "../../components/create/ReqService";
 import AccountTables from "./AccountTables";
-import { useAccountUIStore } from "../../stores/uiStores";
-import { useUserStore } from "../../stores/userStore";
+import { useAccountInit } from "./hooks/useAccountInit";
 
 const Account = () => {
-  const selectedCar = useAccountUIStore((s) => s.selectedCar);
-  const getServicesByIdCar = useUserStore((s) => s.getServicesByIdCar);
-
-  useEffect(() => {
-    getServicesByIdCar(selectedCar?._id);
-  }, [selectedCar, getServicesByIdCar]);
+  useAccountInit();
 
   return (
     <>
