+++ client/src/pages/account/AccountTable.jsx
@@ -1,20 +1,9 @@
-import { useCallback } from "react";
 import Search from "../../components/table/Search";
 import Table from "../../components/table/Table";
-import { useUserStore } from "../../stores/userStore";
-import { useAccountUIStore } from "../../stores/uiStores";
-import useFilteredData from "../../hooks/useFilteredData";
-import { carFilterFn } from "./utils/accountValidation";
-import { handleCarAction as handleCarActionUtil } from "./utils/accountHandlerUtils";
+import { useAccountTableData } from "./hooks/useAccountTableData";
 
 export default function AccountTable() {
-  const user = useUserStore((s) => s.user);
-  const { setSelectedCar, toggleReqService, toggleServices } = useAccountUIStore();
-
-  const memoizedCarFilterFn = useCallback(carFilterFn, []);
-  const { displayData: displayCars, handleSearch } = useFilteredData(user?.cars, memoizedCarFilterFn);
-
-  const handleCar = (e) => handleCarActionUtil(e, user?.cars, setSelectedCar, { toggleReqService, toggleServices });
+  const { displayCars, handleSearch, handleCar } = useAccountTableData();
   const trTh = (
     <tr>
       <th>brand</th>
